+++ 3dc/avp/win95/Frontend/AvP_UserProfile.cpp
@@ -208,19 +208,19 @@ static int LoadUserProfiles(void)
 	}
 
 	// get any path in the load_name
-	int nPathLen = 0;
+	ptrdiff_t nPathLen = 0;
 	const char * pColon = strrchr(load_name,':');
 	if (pColon) nPathLen = pColon - load_name + 1;
 	const char * pBackSlash = strrchr(load_name,'\\');
 	if (pBackSlash)
 	{
-		int nLen = pBackSlash - load_name + 1;
+		ptrdiff_t nLen = pBackSlash - load_name + 1;
 		if (nLen > nPathLen) nPathLen = nLen;
 	}
 	const char * pSlash = strrchr(load_name,'/');
 	if (pSlash)
 	{
-		int nLen = pSlash - load_name + 1;
+		ptrdiff_t nLen = pSlash - load_name + 1;
 		if (nLen > nPathLen) nPathLen = nLen;
 	}
